+++ monorepo/packages/mantine/libs/icons/v0.2.0/icons.ts
@@ -5,7 +5,7 @@ import {
     IconArrowRight, IconFolder, IconPencil, IconLogout2, IconRefresh, IconWand, IconBrush, IconFileTypeXls, IconInfoCircle, IconArchive,
     IconBriefcase2, IconX, IconDownload, IconChevronUp, IconThumbUp, IconThumbDown, IconCircleCheck, IconReportAnalytics, IconRepeat,
     IconSquareNumber1, IconUrgent, IconUsers, IconQrcode, IconMinus, IconSwitchVertical, IconHelp, IconFolderPlus, IconClipboard,
-    IconCubePlus, IconLayersSubtract, IconKey, IconSend, IconArrowLeft, IconArrowBackUp, IconArrowUp, IconArrowsVertical, Icon,
+    IconCubePlus, IconLayersSubtract, IconKey, IconSend, IconArrowLeft, IconArrowBackUp, IconArrowUp, IconArrowsVertical, Icon, IconPrinter,
 } from '@tabler/icons-react'
 
 export default function icons(iconName: string) {
@@ -16,7 +16,7 @@ export default function icons(iconName: string) {
         IconArrowRight, IconFolder, IconPencil, IconLogout2, IconRefresh, IconWand, IconBrush, IconFileTypeXls, IconInfoCircle, IconArchive,
         IconBriefcase2, IconX, IconDownload, IconChevronUp, IconThumbUp, IconThumbDown, IconCircleCheck, IconReportAnalytics, IconRepeat,
         IconSquareNumber1, IconUrgent, IconUsers, IconQrcode, IconMinus, IconSwitchVertical, IconHelp, IconFolderPlus, IconClipboard,
-        IconCubePlus, IconLayersSubtract, IconKey, IconSend, IconArrowLeft, IconArrowBackUp, IconArrowUp, IconArrowsVertical,
+        IconCubePlus, IconLayersSubtract, IconKey, IconSend, IconArrowLeft, IconArrowBackUp, IconArrowUp, IconArrowsVertical, IconPrinter,
     }
     return icons[iconName] as Icon
 }