+++ monorepo/packages/mantine/nodes/elements/dataDisplay/icon/fabricVersions/v0.1.0/IconNode.tsx
@@ -3,6 +3,7 @@ import { getGroupedPorts, getPorts } from '../../../../../../../../libs/nodesFab
 
 import v0_3_0 from './v0.3.0/Icon'
 import v0_4_0 from './v0.4.0/Icon'
+import v0_5_0 from './v0.5.0/IconLazy'
 
 //===================================================================
 
@@ -14,6 +15,13 @@ const compVersions: CompVersions = {
     'v0.4.0': {
         Comp: v0_4_0,
         inputs: [...getGroupedPorts('input', ['Icon']), ...getPorts('input', ['iconColor', 'useScope', 'scope'])]
+    },
+    'v0.5.0': {
+        Comp: v0_5_0,
+        inputs: getPorts('input', [
+            'useScope', 'scope', 'iconType', 'iconName', 'iconSize', 'stroke', 'iconColor', 'themeIconVariant', 'themeIconSize',
+            'themeIconRadius', 'themeIconColor', 'themeIconGradient',
+        ])
     }
 }