+++ monorepo/packages/mantine/nodes/elements/typography/list/fabricVersions/v0.1.0/ListNode.tsx
@@ -1,0 +1,20 @@
+import { getReactNode } from '../../../../../../../../libs/nodesFabric/v0.1.0/getNode/v0.7.0/getNode'
+import { getGroupedPorts, getPorts } from '../../../../../../../../libs/nodesFabric/v0.1.0/portsFabric/v0.5.0/get'
+
+import v0_1_0 from './v0.1.0/ListLazy'
+
+//===================================================================
+
+const compVersions: CompVersions = {
+    'v0.1.0': {
+        Comp: v0_1_0,
+        inputs: [
+            ...getPorts('input', ['listScheme', 'size', 'listType', 'withPadding'], ['listScheme']),
+            ...getGroupedPorts('input', ['Embedded icon']),
+        ]
+    }
+}
+
+//===================================================================
+
+export default getReactNode('List', compVersions, true)