+++ monorepo/packages/pdf/nodes/pdfPage/fabricVersions/v0.1.0/v0.1.0/types.d.ts
@@ -1,0 +1,6 @@
+import ReactPDF from "@react-pdf/renderer"
+import { BaseProps } from "../../../../../../mantine/nodes/types"
+
+export type CompProps = BaseProps & {
+    pdfPageStyles?: ReactPDF.Styles
+}