+++ monorepo/packages/qrCode/nodes/nodesStore.ts
@@ -1,0 +1,3 @@
+import getQrCodeNode_f0_1_0 from "./getQrCode/fabricVersions/v0.1.0/getQrCodeNode"
+
+export default [getQrCodeNode_f0_1_0]